create separate commands folder
