+++ src/helpers.ts
@@ -1,4 +1,4 @@
-export function checkForFloyd(): boolean {
+export function isFloydProject(): boolean {
     // TODO: find a way to check if app is a floyd app
     return true;
 }
